# docs/03_api/openapi.yaml
openapi: 3.1.0
info:
  title: Moonshot Backend
  version: 0.6.0
paths:
...
  /v1/sessions:
    get:
      tags:
      - sessions
      summary: List Sessions
      operationId: list_sessions_v1_sessions_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties:
                  items:
                    $ref: '#/components/schemas/Session'
                  type: array
                type: object
                title: Response List Sessions V1 Sessions Get
      security:
      - HTTPBearer: []
    post:
      tags:
      - sessions
      summary: Create Session
      operationId: create_session_v1_sessions_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
...
  /v1/sessions/{session_id}/sql/run:
    post:
      tags:
      - simulator-runtime
      summary: Run Sql Query
      operationId: run_sql_query_v1_sessions__session_id__sql_run_post
      security:
      - HTTPBearer: []
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Session Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SQLRunRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SQLRunResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/sessions/{session_id}/sql/history:
    get:
      tags:
      - simulator-runtime
      summary: List Sql History
      operationId: list_sql_history_v1_sessions__session_id__sql_history_get
      security:
      - HTTPBearer: []
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Session Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SQLHistoryResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/sessions/{session_id}/python/run:
    post:
      tags:
      - simulator-runtime
      summary: Run Python Code
      operationId: run_python_code_v1_sessions__session_id__python_run_post
      security:
      - HTTPBearer: []
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Session Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PythonRunRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PythonRunResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/sessions/{session_id}/python/history:
    get:
      tags:
      - simulator-runtime
      summary: List Python History
      operationId: list_python_history_v1_sessions__session_id__python_history_get
      security:
      - HTTPBearer: []
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Session Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PythonHistoryResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/sessions/{session_id}/dashboard/state:
    get:
      tags:
      - simulator-runtime
      summary: Get Dashboard State
      operationId: get_dashboard_state_v1_sessions__session_id__dashboard_state_get
      security:
      - HTTPBearer: []
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Session Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardState'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/sessions/{session_id}/dashboard/action:
    post:
      tags:
      - simulator-runtime
      summary: Apply Dashboard Action
      operationId: apply_dashboard_action_v1_sessions__session_id__dashboard_action_post
      security:
      - HTTPBearer: []
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Session Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DashboardActionRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardState'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    DashboardActionRequest:
      properties:
        action_type:
          type: string
          title: Action Type
        payload:
          additionalProperties: true
          type: object
          title: Payload
      type: object
      required:
      - action_type
      title: DashboardActionRequest
    DashboardState:
      properties:
        filters:
          additionalProperties: true
          type: object
          title: Filters
        view:
          type: string
          title: View
          default: default
        annotations:
          items:
            type: string
          type: array
          title: Annotations
      type: object
      title: DashboardState
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PythonHistoryItem:
      properties:
        code:
          type: string
          title: Code
        ok:
          type: boolean
          title: Ok
        stdout:
          anyOf:
          - type: string
          - type: 'null'
          title: Stdout
        stderr:
          anyOf:
          - type: string
          - type: 'null'
          title: Stderr
        plot_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Plot Url
        artifacts:
          items:
            $ref: '#/components/schemas/RuntimeArtifact'
          type: array
          title: Artifacts
        error:
          anyOf:
          - type: string
          - type: 'null'
          title: Error
        runtime_ms:
          type: integer
          title: Runtime Ms
          default: 0
        executed_at:
          type: string
          format: date-time
          title: Executed At
      type: object
      required:
      - code
      - ok
      title: PythonHistoryItem
    PythonHistoryResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/PythonHistoryItem'
          type: array
          title: Items
      type: object
      title: PythonHistoryResponse
    PythonRunRequest:
      properties:
        code:
          type: string
          title: Code
        template_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Template Id
        round_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Round Id
        dataset_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Dataset Id
      type: object
      required:
      - code
      title: PythonRunRequest
    PythonRunResponse:
      properties:
        ok:
          type: boolean
          title: Ok
        stdout:
          anyOf:
          - type: string
          - type: 'null'
          title: Stdout
        stderr:
          anyOf:
          - type: string
          - type: 'null'
          title: Stderr
        plot_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Plot Url
        artifacts:
          items:
            $ref: '#/components/schemas/RuntimeArtifact'
          type: array
          title: Artifacts
        runtime_ms:
          type: integer
          title: Runtime Ms
      type: object
      required:
      - ok
      - runtime_ms
      title: PythonRunResponse
...
    RuntimeArtifact:
      properties:
        name:
          type: string
          title: Name
        mime_type:
          type: string
          title: Mime Type
        uri:
          type: string
          title: Uri
        bytes:
          type: integer
          title: Bytes
        kind:
          type: string
          title: Kind
      type: object
      required:
      - name
      - mime_type
      - uri
      - bytes
      - kind
      title: RuntimeArtifact
...
    SQLHistoryItem:
      properties:
        query:
          type: string
          title: Query
        ok:
          type: boolean
          title: Ok
        row_count:
          anyOf:
          - type: integer
          - type: 'null'
          title: Row Count
        columns:
          items:
            type: string
          type: array
          title: Columns
        error:
          anyOf:
          - type: string
          - type: 'null'
          title: Error
        executed_at:
          type: string
          format: date-time
          title: Executed At
      type: object
      required:
      - query
      - ok
      title: SQLHistoryItem
    SQLHistoryResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/SQLHistoryItem'
          type: array
          title: Items
      type: object
      title: SQLHistoryResponse
    SQLRunRequest:
      properties:
        query:
          type: string
          title: Query
      type: object
      required:
      - query
      title: SQLRunRequest
    SQLRunResponse:
      properties:
        ok:
          type: boolean
          title: Ok
        row_count:
          type: integer
          title: Row Count
        columns:
          items:
            type: string
          type: array
          title: Columns
        rows:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Rows
        runtime_ms:
          type: integer
          title: Runtime Ms
      type: object
      required:
      - ok
      - row_count
      - runtime_ms
      title: SQLRunResponse
...
    Session:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        tenant_id:
          type: string
          title: Tenant Id
        task_family_id:
          type: string
          format: uuid
          title: Task Family Id
        candidate_id:
          type: string
          title: Candidate Id
        status:
          type: string
          title: Status
          default: active
        policy:
          additionalProperties: true
          type: object
          title: Policy
        final_response:
          anyOf:
          - type: string
          - type: 'null'
          title: Final Response
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - tenant_id
      - task_family_id
      - candidate_id
      title: Session
    SessionCreate:
      properties:
        task_family_id:
          type: string
          format: uuid
          title: Task Family Id
        candidate_id:
          type: string
          title: Candidate Id
        policy:
          additionalProperties: true
          type: object
          title: Policy
      type: object
      required:
      - task_family_id
      - candidate_id
      title: SessionCreate
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer